transformed articla to SearchPlace Component
