+++ package.json
@@ -7,7 +7,7 @@
     "node": "6.9.1"
   },
   "scripts": {
-    "test": "npm run jasmine && ava **/*.test.js",
+    "test": "karma start --single-run --browsers PhantomJS",
     "jasmine": "node node_modules/jasmine/bin/jasmine.js",
     "start": "node server/server.js",
     "build": "webpack -d && webpack-dev-server --content-base dist/ --inline --hot --port 3000",
@@ -61,11 +61,19 @@
     "enzyme": "^2.7.0",
     "expect": "^1.20.2",
     "jasmine": "^2.5.3",
+    "jasmine-check": "^0.1.5",
+    "jasmine-core": "^2.5.2",
     "jasmine-enzyme": "^2.1.0",
-    "karma": "^1.4.0",
+    "karma": "^1.4.1",
+    "karma-babel-preprocessor": "^6.0.1",
     "karma-chrome-launcher": "^2.0.0",
     "karma-jasmine": "^1.1.0",
+    "karma-phantomjs-launcher": "^1.0.2",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-webpack": "^2.0.2",
+    "phantomjs-prebuilt": "^2.1.14",
     "react-addons-test-utils": "^15.4.2",
+    "react-test-utils": "0.0.1",
     "webpack-dev-server": "^1.16.2"
   },
   "eslintConfig": {
