+++ src/components/current_plan_list.js
@@ -1,0 +1,8 @@
+import React, { Component } from 'react';
+
+
+class CurrentPlanList extends Component {
+  render() {
+    return null;
+  }
+}
