+++ src/reducers/index.js
@@ -1,7 +1,6 @@
 import { combineReducers } from 'redux';
 import plan from './plan';
 
-// we will import our reducers here & we'll combine it
 
 export default combineReducers({
   plan
